+++ install.sh
@@ -1,12 +1,29 @@
 #!/bin/bash
 
-mkdir -p bin
+# shellcheck disable=SC2164
+# shellcheck disable=SC2103
 
-cat src/*.sh > bin/temp.sh
-cat bashunit >> bin/temp.sh
-grep -v '^source' bin/temp.sh > bin/bashunit
-rm bin/temp.sh
+cd "$(dirname "$0")"
+rm -f lib/bashunit
+[ -d "lib" ] || mkdir lib
+cd lib
 
-chmod +x bin/bashunit
+TAG=${1-main}
 
-echo "Build complete. bashunit has been generated in the bin folder."
+if [[ $TAG == main ]]; then
+  echo "> Using main branch"
+  git clone git@github.com:TypedDevs/bashunit.git temp_bashunit
+else
+  echo "> Using a concrete tag '$TAG'"
+  curl -L -O -J "https://github.com/TypedDevs/bashunit/archive/refs/tags/$TAG.tar.gz"
+  tar -zxvf "bashunit-$TAG.tar.gz"
+  cp "bashunit-$TAG/bin/bashunit" temp_bashunit
+  rm "bashunit-$TAG.tar.gz"
+fi
+
+## Common
+cd temp_bashunit
+./build.sh
+cd ..
+cp temp_bashunit/bin/bashunit bashunit
+rm -rf temp_bashunit